requirements.txt
fastapi              #==0.114.2
uvicorn[standard]    #==0.30.6
python-multipart     #==0.0.9
pymongo              #==4.8.0
httpx                #==0.27.2
python-docx          #==1.1.2
PyMuPDF              #==1.24.10
pillow               #==10.4.0
transformers         #==4.44.2
torch                #==2.4.0
sentence-transformers#==3.1.1
sumy                 #==0.11.0
numpy                #==1.26.4a